remove unnecessary code from utils
